add more error checking and tests for the checks

+++ ext/utf8/utf8.h
@@ -2,6 +2,6 @@
 #define UTF8_UTF8_H
 
 inline int8_t utf8CharLen(unsigned char *in, size_t in_len);
-size_t utf8CharCount(unsigned char *in, size_t in_len);
+int64_t utf8CharCount(unsigned char *in, size_t in_len);
 
 #endif
